Fix .xdc creation example
The described process creates an invalid .xdc file.  if i have
`index.html` in a directory `foo/`, and the previous command was run,
the zipfile would contain `foo/index.html`.  But webXDC requires plain
`index.html` in the zipfile.

The updated command creates the expected webxdc file structure (tested
with zip 3.0-13 on debian)